More 'make check' warning fixes.
git-svn-id: svn://svn.valgrind.org/valgrind/trunk@9239

+++ none/tests/rlimit_nofile.c
@@ -44,8 +44,9 @@ int main(int argc, char **argv)
 
    if (newrlim.rlim_max != oldrlim.rlim_max)
    {
-      fprintf(stderr, "rlim_max is %lu (should be %lu)\n",
-              newrlim.rlim_max, oldrlim.rlim_max);
+      fprintf(stderr, "rlim_max is %llu (should be %llu)\n",
+              (unsigned long long)newrlim.rlim_max,
+              (unsigned long long)oldrlim.rlim_max);
    }
 
    newrlim.rlim_cur -= 3; /* allow for stdin, stdout and stderr */
